requirements.txt
Flask
Flask-CORS
python-dotenv
gunicorn
pandas
requests
